% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/womeninventoR-package.R
\docType{package}
\name{womeninventoR-package}
\alias{womeninventoR}
\alias{womeninventoR-package}
\title{womeninventoR: This package loads the womeninventorpatents dataset as a tibble that is ready for analysis within R.}
\description{
This package shares patent data from Women inventors obtained from USPTO's Patentsview page here: https://patentsview.org/download/data-download-table
}
\author{
\strong{Maintainer}: Enric Escorsa \email{enricescorsa@gmail.com} (\href{https://orcid.org/0000-0001-8411-6075}{ORCID})

}
\keyword{internal}
